Various quick hacks to get Concordion running on Android.
Currently it searches for HTML specs in the "assets" directory of the
test project, and writes output to /data/local/tmp, which is normally
world writable - rather than having to add SD permissions to the app.

However, this alone is not enough for Concordion to run properly on
Android devices, as various Xerces and java.bean classes are missing.

More investigation is needed to determine whether it will be possible
to run Concordion properly on a standard Android installation without
having to do some major surgery.